requirements.txt
requests==2.31.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==6.0.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.16.0
